shill: RDEPEND on Floss
The Floss ebuild installs btmanagerd, which shill needs to query over
D-Bus to collect information about the state of BT (is it enabled, does
it use Floss or Bluez, ...)
Add the corresponding RDEPEND to shill's ebuild.

Also add the 'floss' USE flag to the ebuild to avoid installing Floss on
some boards for now. If the 'floss' USE flag is not enabled, shill won't
compile in support for communicating with btmanagerd over D-Bus.

BUG=b:253297825
TEST=cros-workon-${BOARD} start shill
TEST=emerge-${BOARD} shill
TEST=equery-${BOARD} g shill |grep floss
TEST=USE="-floss" emerge-${BOARD} shill
TEST=USE="-floss" equery-${BOARD} g shill |grep floss

Change-Id: Ifb1b2f3e91d9f06b86bbb1fd6084b038ef6af9ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4103923
Reviewed-by: David Ruth <druth@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>