llvm: bundle new llvm next flag
This CL bundles changes from the compiler wrapper which add:
    -Wno-single-bit-bitfield-constant-conversion
    -Wno-array-parameter
to the compiler wrapper for LLVM Next.

These flag exists in LLVM Next, but do not yet exist in LLVM Current.
These new warnings were causing a breakage when testing the LLVM uprev.

This CL is a noop for CQ and current LLVM users.

BUG=b:262409607, b:262076232
TEST=bundle.py

Change-Id: Ifbc0878bf5bcc3573b3138117e42a038ef9985aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4104063
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>