ml-test-assets: update model zoo with new models, move to google storage
This CL fixes the convolution benchmarks that contained NaN's in some of
the weights.

This also introduces the mobilenet models are a proxy benchmark for
image search.

This also moves the storage of the model files out of git and into
google storage.

BUG=b:263313506
TEST=None

Cq-Depend: chromium:4113556
Change-Id: Icf3f2bfd2da2d07c7f1fa91b3c509d3fe26b462f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4113504
Tested-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Michael Olney <mpolney@google.com>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>