Rust upgrade to 1.66.0.
Two things worth noting:

rust-bootstrap-use-CARGO_HOME.patch

is new. We need this to build using RAW_SOURCES due to a change that I
suspect is a bug in how bootstrap.py uses CARGO_HOME. See this issue:

https://github.com/rust-lang/rust/issues/106019

Also, in cros-config.toml, we now set
`optimized-compiler-builtins = true`
or we get linker errors. This is a new option.

BUG=b:262876874
TEST=CQ

Change-Id: I1bd812c1135e0d492ed5d6c376729cf3c63b740a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4121236
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Michael Benfield <mbenfield@google.com>