autotest-server-tests-shill: Maintain parity between tauto wifi tests
Our autotest-server-tests-shill ebuild has not been maintained to
maintain parity between the wifi tests in tree in our autotest
directory. The behavior of having a mismatch between a test in a suite
and not being present in an ebuild is not tolerable for CFT (which our
tests have been migrated to) but was permissible with the legacy system.

To address this issue, we update the ebuild to match the wifi autotests
in our tree.

BUG=b:263168742
TEST=cq

Change-Id: I2e034715ac3077359af8cdc00e32d3ce2a24a93b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4121475
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>