cros-kernel2: Add ability to specifically choose dt files
Bundling every device tree under the sun with our kernel is a
waste. Let's add an ability to limit which device trees are being
included. Boards can specify a glob with the variable "CHROMEOS_DTBS"
and that will limit things. We'll also add a USE flag to get back the
old behavior in case someone (like me) wants a slightly more universal
kernel.

Overlays will need to be updated to specify the proper glob to take
advantage of this.

BUG=b:218406702, b:263310098
TEST=Build trogdor/herobrine w/ and w/out the USE flag
TEST=Build an overlay that doesn't specify CHROMEOS_DTBS

Change-Id: Ie055fe287bb2e87de85b0d6ec7410046e50dbff3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4120704
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>