cros-kernel2: Stop arbitrarily listing the first device tree as default
Since the dawn of the use of FIT images in ChromeOS, we've always
listed the "first" device tree as the default that the bootloader
should use if it can't find a compatible that matches. This just
doesn't make any sense. If we can't find any compatible device tree,
we shouldn't just pick one and hope for the best. In the best case,
picking the wrong device tree could cause the device to crash in a
confusing way where it takes some time for a developer to figure out
that they simply forgot to bundle the right device tree. In the worst
case, it could cause a device to crash and burn (literally, since in
theory it could get massively unlucky and overvolt some regulator).

Let's just remove the "default" specifier. Now, depthcharge will
report a message like this if it can't find a matching device tree:

    No compatible or default configs. Giving up.

That's much better.

BUG=b:218406702, b:263310098
TEST=Try booting a device that doesn't have its device tree bundled

Change-Id: Ib943958fc6def0c022d973bc92bf19a672f52ed0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4120703
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Julius Werner <jwerner@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>
Reviewed-by: Alexandru Stan <amstan@chromium.org>