chromeos-base/crosvm: add trace-marker use flag
With https://crrev.com/c/4075413 we have added a new feature to enable
trace_marker tracing with cros_tracing. We need to expose the option to
emerge crosvm with this feature flag enabled. It is disabled by
default as developers are expected to build crosvm locally with the flag
if they need to.

BUG=b:256052372
TEST=`USE="$USE crosvm-trace-marker" emerge-$BOARD crosvm` and deployed

Change-Id: I5f5c34a26bdcae489e3965a556096a0df494e839
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4123791
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Auto-Submit: Morg <morg@chromium.org>
Tested-by: Morg <morg@chromium.org>