ffmpeg: Add resolution-without-keyframe patch
This CL adds a downstream patches [1] to allow for ffmpeg decoding
the VP9 test vectors ending up in (frm_resize|sub8x8_sf). This is
a temporary solution and continuation of the upstream series [2];
since ffmpeg is a test-only binary on chromeOS and this test has
been failing since the beginning, this CL just takes that patch [1]
as-is.

[1] https://github.com/intel-media-ci/ffmpeg/pull/608
[2] https://patchwork.ffmpeg.org/project/ffmpeg/list/?series=7915

BUG=b:234847837
TEST=video.PlatformDecoding.ffmpeg_vaapi* on volteer

Change-Id: Ia5733bf8c4c2591a640ae1d92e7457a7866b22ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4061439
Tested-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@google.com>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>