libchromeos-use-flags: Disable Floss for boards that do not support it.
Currently, if the Floss flag is enabled and Floss does not exist on a
given 2GB rootfs board, Bluetooth is disabled entirely.

To fix this, we prevent users from enabling Floss in the first place
by override enabling/disabling Floss in the Chromium UI if the USE
flag is present/absent. The presence of the USE flag is indicated by
an explicit command line disable when Chrome is built.

BUG=b:228902194
TEST=manual

Cq-Depend: chromium:4127877
Change-Id: I008820c1502931e736830eb9120c91a036f51a4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4128738
Tested-by: Sarvesh Kalwit <sarveshkalwit@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>