skyrim: Add markarth coreboot configuration
Create a new coreboot configuration for the markarth variant
of the skyrim reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:262092858
TEST=FW_NAME=markarth emerge-skyrim coreboot chromeos-bootimage
Ensure that image-markarth.*.bin are created

Change-Id: I09ea1645bd75dece25a2540e1a3279bb550a9b47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4131017
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Tested-by: Chao Gui <chaogui@google.com>
Commit-Queue: Chao Gui <chaogui@google.com>