sys-firmware: add edk2-ovmf-crosvm ebuild
This is a patched version of the EDK2 OVMF (Open Virtual Machine
Firmware) that is compatible with the crosvm hypervisor. It will be used
to build a BIOS image that can boot UEFI-compatible bootloaders and
operating systems inside the VM.

Based on the upstream Gentoo sys-firmware/edk2-ovmf ebuild, with
modifications to apply crosvm-specific patches to work around missing
functionality and differences between crosvm, QEMU, and CloudHypervisor.
Over time, these patches should be removed as crosvm is improved or
upstream EDK2 is modified to support crosvm as a hypervisor platform.

This version of the ebuild uses the edk2-stable202211 tag from the main
edk2 repository.

BUG=b:264495396
TEST=Boot to UEFI shell inside crosvm

Change-Id: I46ed4aaee4002aad08dae2880d927d37da722782
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4152885
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>