cros-kernel2: Disable CONFIG_FORTIFY_SOURCE in build tests
CONFIG_FORTIFY_SOURCE results in inexplicable build errors, such as

include/linux/fortify-string.h:543:4: error:
    call to __read_overflow declared with 'error' attribute: detected read beyond size of object (1st parameter)
                        __read_overflow();

when trying to build chromeos-6.1 with the compiler used in chroot.
The problem is not seen with gcc or with clang version 14.0.

There is nothing obviously wrong with the code, meaning the compiler
is acting up. Disable CONFIG_FORTIFY_SOURCE in build tests to avoid
the error.

BUG=b:265129193
TEST=USE=buildtest emerge-arm-generic chromeos-kernel-6_1

Change-Id: Id6c2afe7ae15b2456a86c6e1e889c4c7fd8b7c20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4154399
Tested-by: Tzung-Bi Shih <tzungbi@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>