chromeos-base/ec-utils: Enable manual uprev
Configure manual uprev for the packages "chromeos-base/ec-devutils",
"chromeos-base/ec-utils", and "chromeos-base/ec-utils-test" and document
how to manually uprev in case of emergency.

Since the CQ for platform/ec doesn't normally build ebuilds, if there is
a breakage, it should be caught by the PUpr commit instead of breaking
everyone.

BUG=b:261258378
TEST=Ran cros_mark_as_stable commands

Change-Id: I8cd90a797f7ef77796b2f66650a1ec1d87aa766e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4157987
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>