sys-kernel/linux-headers: Update AV1 stateless decoder header to v4
The AV1 stateless V4L2 decoder uAPI has been updated. Changes include:

- V4L2_AV1_NUM_REF_FRAMES macro dropped
- struct v4l2_av1_film_grain name changed to v4l2_ctrl_av1_film_grain
- v4l2_av1_global_motion.params[] type changed from __u32 to __s32
- v4l2_ctrl_av1_frame.ref_frame_idx type changed from __u8 to __s8
- v4l2_ctrl_av1_frame.last_frame_idx and
  v4l2_ctrl_av1_frame.gold_frame_idx fields removed

Update our copy so that we can follow and participate in upstream
development.

Most of the changes will break the Chromium v4l2_stateless_decoder test
application build. The required changes on that side are in a separate
patch in a separate repo. Landing this and the other patch need to be
planned.

BUG=b:265063647
TEST=emerge sys-kernel/linux-headers

Change-Id: I51acb543244726d6b0d73adaa89a47e73ac97eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4167050
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Chen-Yu Tsai <wenst@chromium.org>
Commit-Queue: Steve Cho <stevecho@chromium.org>