chrome-lacros: Make lacros-mounter use session namespace if requested
Guest sessions use their own namespace. We need to mount Lacros
there in order for Ash to be able to launch it. Introduce a parameter to
the init script to do so. It will be passed by Chrome's BrowserManager
in the future. Until the necessary change in Chrome happens, the
overall behavior is unchanged.

BUG=chromium:1347937
TEST=manually

Change-Id: Iaeae5b2da6cfc8a90f2e9d6a95f678af981bb8d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4179664
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: Georg Neis <neis@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Georg Neis <neis@chromium.org>