chromeos-base: Add GuestOS team as termina-dlc owners
Currently this is only owned by the uprev service account, presumably
since we don't make non-automated changes often and whoever set up the
owners didn't realise we should be on it.

I'm using the OWNERs file in container-guest-tools as the canonical
list of team members, since that repo also contains a bunch of
configuration and packages for termina and the two components should
be co-owned.

BUG=none
TEST=none

Change-Id: Ie51c83f9e26b7e540b80d221feb2f42a7c9368fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4183896
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>