cups: Enable extra debug messages
cupsd has a bunch of debugging messages that are not compiled in by
default, so they can't be turned on even by changing the debug level.
In the past this didn't matter much because users couldn't change the
log level anyway.  Now that we have the ability to change the log level
on the fly, it's useful to compile these in so that we can turn them on
with printscan_debug.  Messages guarded by this flag will still not be
emitted unless printer debug logging is enabled.

Messages are emitted to stderr, which is discarded by default.  To
capture these and redirect them into syslog, wrap the cupsd minijail in
a syslog-cat call.

BUG=b:192017863
TEST=printscan_debug start printing; lpstat -v

Cq-Depend: chromium:4209576
Change-Id: I28aa41ea5a9403319fd8c7a415cb0734e937daff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4206359
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Mason Wilde <masonwilde@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>