brya: Add constitution coreboot configuration
Create a new coreboot configuration for the constitution variant
of the brask reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:267539938
TEST=FW_NAME=constitution emerge-brya coreboot chromeos-bootimage
Ensure that image-constitution.*.bin are created

Change-Id: I1ddf1bfcc43830cfe02449e6b93b27550723b3f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4216058
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: Pablo Ceballos <pceballos@chromium.org>
Commit-Queue: David Wu <david_wu@quanta.corp-partner.google.com>
Reviewed-by: Matthew Ziegelbaum <ziegs@chromium.org>
Tested-by: David Wu <david_wu@quanta.corp-partner.google.com>