brask: Add aurash coreboot configuration
Create a new coreboot configuration for the aurash variant
of the brask reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:263691099
TEST=FW_NAME=aurash emerge-brask coreboot chromeos-bootimage
Ensure that image-aurash.*.bin are created

Cq-Depend: chromium:4236659
Change-Id: I4c2ade0a016091c457d54c84f906042b422d1a7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4236491
Reviewed-by: Ricky Chang <rickytlchang@chromium.org>
Reviewed-by: Zoey Wu <zoey_wu@wistron.corp-partner.google.com>
Tested-by: Zoey Wu <zoey_wu@wistron.corp-partner.google.com>
Commit-Queue: Zoey Wu <zoey_wu@wistron.corp-partner.google.com>