dev-cpp/gtest: Build googletest from manifest checkout
Firmware needs to build googletest from source, so googletest is now
synced down through the repo manifest. We want the ebuild to remain in
sync with the version checked out (to avoid having multiple different
versions of googletest in ChromeOS), so this commit changes the ebuild
to use the "cros-workon" framework.

BUG=b:254530679
TEST=CQ passes

Cq-Depend: chrome-internal:5450525
Change-Id: Ic26d71727dc9e10cfbae475d505e3a4be9c099c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4241336
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>