sys-devel: Mark latest patch as chromimumos
This CL is a no-op for ChromeOS, and does not affect builds, tooling,
or LLVM. Because of this, LLVM is not getting a revbump.

There's a typo here. "chromeos" isn't recognised as a valid platform,
as the patches apply to "chromiumos". This unblocks the patch sync.

BUG=b:268657960
TEST=CQ

Change-Id: I988dc78954101f2e412921db36a2836c97e1b80a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4242023
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Auto-Submit: Jordan Abrahams-Whitehead <ajordanr@google.com>