net-misc: modemmanager-next: use builtin plugins exclusively
Do not install the plugins as separate files that are dlopen()-ed on
runtime, default to have them built in the daemon binary itself.

This makes the boot time of the daemon much quicker, and also makes it
more secure as it avoids dynamically loaded modules.

There is no change in the behavior of the daemon, apart from the
shorter boot time as it doesn't need to load plugins one by one.

BUG=None
TEST=CQ passes

Change-Id: I33c42f5cc832f1df56fbae89859bb25ed1204784
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4272863
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>