eclass: cros-workon: Search for pre-created tree artifacts
In order to support building pinned cros-workon packages with bazel,
we need to inject the pinned src into the container. We could inject the
.git repo and mount it in /mnt/host/sources/src, but that poses two
problems:
1) There is no hermetic format for a git repository, so we would
   constantly cache bust between users and any `repo sync` operation.
2) This results in an extra copy operation. cros-workon will `checkout`
   the pinned src from the /mnt/host/sources/src .git repository into
   the ebuild's work directory.

This change introduces a new `trees` directory that we can pre-populate
with a hermetic artifact that is generated from the tree hash.
cros-workon will check this directory when it's constructing the workdir
for the ebuild. By having cros-workon perform the extraction we can
simply bind mount the `tarball` into the container and avoid the extra
step of populating /mnt/host/sources/src.

This change will be a no-op for the portage builds because the `trees`
cache directory will only be populated by bazel.

BUG=b:269606353, b:262460834
TEST=BOARD=skyrim bazel run @portage//sys-boot/libpayload:0.0.1-r4296_debug -- --login=after
TEST=BOARD=skyrim bazel run @portage//chromeos-base/chromeos-config:0.0.2-r194

(cherry picked from commit ca50739f30372b8984308168a334167e82a906d9)

Original-Change-Id: I0bdda9cdb21e079d23113a4e05de0953b22dcccd
Original-Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/third_party/chromiumos-overlay/+/1753797
Original-Reviewed-by: Shuhei Takahashi <nya@google.com>
Change-Id: I0bdda9cdb21e079d23113a4e05de0953b22dcccd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4312392
Reviewed-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>