libhwsec: limit job count when tests are enabled
It's strongly suspected that this ebuild is spawning $(ncpu) Clang
instances that each consume 1-2GB of RAM for a prolonged time on
postsubmit bots. Since $(ncpu) == 32 and each bot has 64GB of RAM, this
contributes significantly to us hitting memory limits, which fails
builds.

At this point, it's not clear if this is a compiler issue or just
gmock/gtest generating a lot of code to work with. I have more analysis
on b/261541399#comment62 and comment 63.

BUG=b:266352772, b:261541399
TEST=`FEATURES=test emerge libhwsec` used way less peak memory locally

Change-Id: Iaf7855a2e48d1eb969b4092d1064322fe52f4156
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4289893
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>