camera: remove camera_feature_effects guard on ml-core dependency
We need the /var/lib/ml_core/opencl_cache directory to exist in order
for the minijail to be correctly set up for the camera stack. So push
the use of the camera_feature_effects guard into ml-core, which allows
us to set up the tmpfiles.d directory correctly.

BUG=b:257160623
TEST=emerge, deploy and test on brya

Change-Id: I13a181f51618cb9ec7fe5aa90d47a758fa49f3ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4297313
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Tested-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>