chromeos-base/minijail: Drop DISTUTILS_SINGLE_IMPL
CL:1915163 introduces DISTUTILS_SINGLE_IMPL for seemingly no reason.
This is problematic, as it means we'll only install this package for
Python 3.6, and not 3.8.  This means, when we change /usr/bin/python3
to point to 3.8, that scripts which import minijail will stop working.

By removing DISTUTILS_SINGLE_IMPL, this package will now work for both
3.6 and 3.8, allowing us to more safely change /usr/bin/python3.

As a follow-up, I think it would be nice to move the Python from this
package to a separate package, e.g., dev-python/minijail, that way we
can clean up the manual calling of distutils phase functions.

BUG=b:187794810
TEST=python3.8 -c "import minijail"
     compile_seccomp_policy --help
     generate_seccomp_policy --help
     generate_constants_json --help

Change-Id: I436cb603f06525a07a43ebcc1edf9ca99481feee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4294183
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>