cros-rustc: force codegen-units to 32
b/271569975 and b/271306977 have a decent amount of investigation that
tries to determine why artifacts built by two Rust toolchains with
identical sources end up being incompatible. In this case, one Rust
toolchain is built on the SDK builder, and the other is built on my
workstation.

As it turns out, codegen-units ends up being part of cargo's build
configuration, which is included in the hash of per-crate metadata. This
hash is included in yet another hash. The "yet another hash" determines
rlib compatibility. (I'm ignoring extra complexity here, but the point
I'm trying to make isn't impacted by that).

The value of `0` here means "pick a number based on how many cores the
current machine has."

So today, `rustc`'s intermediate build artifacts can only be used by you
if the machine your `rustc` was built on had the same number of cores as
the machine the other `rustc` was built on.

This patch forces everything to 32, since that's the SDK builder's
current value. Doing so removes this core count problem entirely.

BUG=b:271569975
TEST=A lot of builds. A lot. Locally and remotely.

Change-Id: If66f8c4e97ed5add667891aa1ed3d96375b523b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4310013
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>