eclass: cros-workon: Use CROS_WORKON_COMMIT if its a SHA
When invoking an ebuild via bazel, we don't push the .git directory
into the container because it's non-hermetic input. This means we can't
query the repository for the current hash. This change makes it so that
if the CROS_WORKON_COMMIT matches the regex for a SHA1 we directly use
it to set the VCSID.

BUG=b:269606353, b:262460834
TEST=print out the VCSID on both a pinned cros-workon ebuild and a non
pinned version.

Change-Id: I45d56aa05e071ed1b007fd184930d5fb2f18e14b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4312393
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>