sys-devel/gdb: Build without -march flags for amd64
gdb needs to run on both the host and test device.  This means we need
to avoid using any CPU instructions that won't run on the host.  The
result is potentially slightly slower when running on a device, but
that's ok because gdb is only used on test devices and isn't highly
performance sensitive.

BUG=b:272282321
TEST=gdb-grunt on workstation; gdb on grunt dut

Change-Id: I3513817119e64d7ab1bdce09da356f7c32912e84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4320745
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>