sys-apps: Add fwupd-efi
fwupd-efi is needed for enabling UEFI support in fwupd. The imported
version is 1.4, which adds NX compatibility. Note that this version is
currently marked unstable in Gentoo.

This is a pristine copy made with cros_portage_upgrade, except for
changing PYTHON_COMPAT from 9..11 to 6..9 so that emerge doesn't
complain. The ebuild is masked for now since additional fixes are
needed.

BUG=b:244312163
TEST=`emerge-reven --search fwupd-efi` shows ebuild is masked

Change-Id: Ifbd4c2d834e87263b303d191f8c5cb7c86376652
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4330396
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>