libqmi: disable compat layer
The compat layer in libqmi is used to avoid API/ABI breaks during
library updates, and it includes symbols that should no longer be used
in newly written code.

In ChromeOS the library is always built from source along with the
rest of the system, so there is no real need to keep the compat
layer. The library may break API/ABI at any point without affecting
the rest of the system.

BUG=b:257108531
TEST=emerge-$BOARD net-libs/libqmi

Change-Id: I4b1a46a30156c547ce9a15e15e10aa1017c5d5fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4338338
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>