cras-client: don't use "${S}" from pkg_* functions
When installing cras-client as a binpkg, "${S}" isn't available, which
leads to `cros-rust_get_crate_version` returning nothing.

This is the minimal change we can make to fix the CQ. Verification
should be added in a later CL so this is more resilient to changes.

BUG=b:273483838
TEST=`emerge-brya cras-client` still DTRT

Change-Id: Ib769af5869b5770c3cd1ecb454bb25ef11d30c9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4338203
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>