arc-explicit-sync: Move to chromiumos-overlay
We used to put this package in the chipset overlays to for the ARC++
fence sync timeout issue, but this doesn't make much sense because all
it does is to install a rc file with nothing secret inside.

Move this to the common overlay alongside the EAPI upgrade.

BUG=b:273778983
TEST=# Geralt uses ARCVM, so no file installed
     emerge-gearlt -av arc-explicit-sync
     # Kukui still uses ARC++ atm, file is installed
     emerge-kukui -av arc-explicit-sync

Change-Id: Iad993ae094595c5f164c92d7fb419e6c3bf274be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4344440
Commit-Queue: Fei Shao <fshao@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>