nissa: Add yavilla coreboot configuration
Create a new coreboot configuration for the yavilla variant
of the nereid reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:273791621
TEST=FW_NAME=yavilla emerge-nissa coreboot chromeos-bootimage
Ensure that image-yavilla.*.bin are created

Cq-Depend: chrome-internal:5644509, chrome-internal:5778730
Change-Id: I18741cce9b117d3a2ec836517e53359962a6cde2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4344444
Tested-by: Tony Huang <tony-huang@quanta.corp-partner.google.com>
Commit-Queue: Tony Huang <tony-huang@quanta.corp-partner.google.com>
Reviewed-by: Derek Huang <derekhuang@google.com>