linux-firmware: Change flag for AX203 WiFi
The cq test |TestIwlwifiFirmwareAndKernelMatch| assumes the flag name is
the prefix for the firmware name (iwlwifi-so-a0-hr-b0-79.ucode). The
current prefix "iwlwifi-ax203" was chosen because there is already a
similar flag "iwlwifi-so" for a different wifi chipset, and this flag
name fails this cq test. Here, change into a longer prefix as the flag
name to differentiate these two chipsets and make this cq test case
pass.

BUG=b:268116728,b:272819642
TEST=cros_workon --board=nissa start linux-firmware
TEST=USE="linux_firmware_iwlwifi-so-a0-hr" emerge-nissa linux-firmware
TEST=ls /build/nissa/lib/firmware
  make sure iwlwifi-so-a0-hr-b0-79.ucode is there

Change-Id: I8a3e47689231b7f075cac34aa5b56243ba22e1ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4355501
Tested-by: Yichen Yu <yichenyu@google.com>
Reviewed-by: David Ruth <druth@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Yichen Yu <yichenyu@google.com>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Auto-Submit: Yichen Yu <yichenyu@google.com>