virtual/target-chromiumos-sdk: Add missing host SDK tools
We are missing some tools from the SDK tarball. For the bazel build we
need to manually list out the `gs://` urls for each of these prebuilds.
It would be better to have them included in the SDK so we can drop our
hard coded list of additional packages to install.

This change doesn't add hps-sdk, coreboot-sdk or dev-lang/rust since
those are declares in the `cross` virtual.

BUG=b:262460834
TEST=cros_workon --host start target-chromium-os-sdk
TEST=sudo emerge target-chromium-os-sdk

Change-Id: I57d62fcc926d5480e9e923112b90a8742e5404b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4358983
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>