Use drmModeAddFB for DRM_FORMAT_MOD_INVALID.
CL:4104308 added the ability to pass modifiers from waffle to DRM/KMS
when creating a framebuffer. Unfortunately, this is causing
graphics.Smoke.platform failures on AMD Stoney Ridge. The root cause
seems to be that drmModeAddFB2WithModifiers() fails.

This CL updates the patch introduced in that CL with the following
fixes:

1) Use drmModeAddFB2WithModifiers() only if the modifier is not
   DRM_FORMAT_MOD_INVALID. Otherwise, use drmModeAddFB(). This is enough
   to fix the AMD Stoney Ridge failures.

2) Fall back to drmModeAddFB() if drmModeAddFB2WithModifiers() failed.
   This is based on a suggestion by marcheu@ (b/270727667#comment13).

BUG=b:270727667
TEST=graphics.Smoke.platform on grunt

Change-Id: Id1178b158ffa84e9157a41ea24a48c516f40009d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4367383
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Matt Turner <msturner@google.com>