eclass/autotest: Add missing DEPEND
The autotest eclass invokes the `autotest_client` python script as part
of src_compile so it can invoke the `setup()` method for each test. It
also invokes the autotest `packager.py` script as part of
`pkg_postinst`. In theory `autotest` should be a BDEPEND and an IDEPEND,
but we can't even install `autotest` as a BDEPEND right now. One of its
dependencies fails to compile. Given the way the ebuild is written, we
need to declare autotest as a DEPEND so that the python scripts are
available at build time.

Original error that found the problem:

```
 * Compiling tests
 * Building tests (1):
 * myfaketest/build/arm64-generic/tmp/portage/chromeos-base/chromeos-chrome-9999/temp/environment: line 1667: client/bin/autotest_client: No such file or directory
 * ERROR: chromeos-base/chromeos-chrome-9999::chromiumos failed (compile phase):
 ```

BUG=b:262460834
TEST=BOARD=arm64-generic bazel build --//bazel/sdk:new-sdk @portage//chromeos-base/chromeos-chrome
TEST=emerge-arm64-generic -atvj chromeos-base/chromeos-chrome

Change-Id: I597c461ab09553bf009a38278c68dc863c7fde31
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4383109
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>