sys-libs/{libcxx,llvm-libunwind}: Add glibc as a DEPEND
These are primordial packages that we use to bootstrap the system. We
need to add the glibc dependency so that the packages get built in the
correct order.

This should be a nop for the portage build since glibc is added
into package.provided by setup_board.

I'm not sure how the cross- packages work. Do we need to have glibc
listed there?

BUG=b:266979761
TEST=BOARD=arm64-generic bazel build -k --config=prebuilts/arm64-generic  --//bazel/sdk:new-sdk @portage//virtual/target-os:all

Change-Id: I270dad362452549d976b06cacb990c5033b8256f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4384865
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>