bootchart: Collect dmesg + other fixes
pybootchart knows how to parse some pieces of the kernel log if it's
found in the tarball as a 'dmesg' file. Let's collect it.

ARM64 is "aarch64"; fix the uname cases to match that, since we get
slightly better CPU info from the ARM case.

Fix (or mask) lint/shellcheck issues while at it.

BUG=none
TEST=platform.BootPerf with initcall_debug; pybootchart parses kernel
     init stages

Change-Id: Ida6c609bf53ec9b86d07d3aa805a535c0f7b5603
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4402310
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Chinglin Yu <chinglinyu@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>