dev-root: when pvs is installed, disable ssh-server
when the PVS packages are installed on the machine, we'd like to turn
off the ssh server on the test image. This ensures that when PVS
executes and creates test reports on the machine, these test reports are
not accessible from outside of the machine.

BUG=b:257971808
TEST=build image, observe ssh-server not running

Change-Id: I61c183111ccbb71a26e4441faae9e19b1877bc2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4412066
Tested-by: Parker Holloway <jparkerh@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Brett Brotherton <bbrotherton@google.com>
Commit-Queue: Parker Holloway <jparkerh@google.com>