dedede: Add taranza coreboot configuration
Create a new coreboot configuration for the taranza variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:277664211
TEST=FW_NAME=taranza emerge-dedede coreboot chromeos-bootimage
Ensure that image-taranza.*.bin are created

Change-Id: Ia568e85ca42238f3ca33cbf9913d71ba849bb7b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4410614
Tested-by: David Wu <david_wu@quanta.corp-partner.google.com>
Reviewed-by: Sheng-Liang Pan <sheng-liang.pan@quanta.corp-partner.google.com>
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>
Commit-Queue: David Wu <david_wu@quanta.corp-partner.google.com>