rex: Add screebo coreboot configuration
Create a new coreboot configuration for the screebo variant
of the rex0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:276814951
TEST=FW_NAME=screebo emerge-rex coreboot chromeos-bootimage
Ensure that image-screebo.*.bin are created

Change-Id: I2bdb00513ef76a0380e53d608dfa2c0cdc9c8e55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4417658
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Kapil Porwal <kapilporwal@chromium.org>
Tested-by: Simon Zhou <zhouguohui@huaqin.corp-partner.google.com>
Commit-Queue: Kapil Porwal <kapilporwal@chromium.org>