chromeos-base: sysctl: Turn on kernel.hung_task_all_cpu_backtrace
On kernel 5.8+ we will no longer get backtraces of all running CPUs
when a hung task happens. The argument in the commit message for these
traces to be off includes:
* If someone is collecting a full crash they'll have this info
  anyway. We're not.
* If someone is on a system with a boatload of CPUs then this can be a
  lot of info. ChromeOS devices don't have so many CPUs that this is a
  problem.

Those backtraces can be useful in debugging crashes. Let's turn these
traces back on.

NOTE: Even though this property isn't present in older kernels it's
not a problem to specify it in this file. This file is processed by
`chromeos_startup` which calls `sysctl -q --system`. It can be see
that the `sysctl` command silently ignores any properties that it's
unaware of.

BUG=b:278594093
TEST=echo HUNG_TASK > /sys/kernel/debug/provoke-crash/DIRECT

Change-Id: Ifefde00d63977944c2fc3789ed412b3a8bd480bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4436734
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>