dedede: Add boxy coreboot configuration
Create a new coreboot configuration for the boxy variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:277529068
TEST=FW_NAME=boxy emerge-dedede coreboot chromeos-bootimage
Ensure that image-boxy.*.bin are created

Change-Id: I0aca4407c69e7e9fe05fa84c231e37af621c464d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4446196
Tested-by: Kevin3 Yang <kevin3.yang@lcfc.corp-partner.google.com>
Reviewed-by: Derek Huang <derekhuang@google.com>
Commit-Queue: Kevin3 Yang <kevin3.yang@lcfc.corp-partner.google.com>