camera: Fix manipulator test segmentation issue
When a two consecutive frames are sufficently different, the segmentation
mask requires a sufficiently high iterations to blend from one mask to the
next. This causes the expected result to differ as it is still
transitioning to the expected result.

Updating sample image to the same image such that the test produce
correct result. This avoids us to use a very high number of iterations
slowing the test down.

BUG=b:278821053
TEST=cros_effects_sm_tests

Cq-Depend: chromium:4446137
Change-Id: Id82f442bb3ffb5e17a17ea88e159badfa6e41575
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4446197
Tested-by: Jingyuan Tu <zactu@google.com>
Commit-Queue: Jingyuan Tu <zactu@google.com>
Reviewed-by: Jim Pollock <jmpollock@chromium.org>