avahi-daemon: Increase cache size
It is desirable to keep avahi's cache small because some query packets
require an expensive traversal of all cache entries.  However, avahi
doesn't resolve mDNS names reliably when its cache is full, so the cache
needs to be big enough to avoid filling up under normal use.  Depending
on the network setup and the services offered by a device, an individual
device can take up anywhere from 4-60 entries in the cache, so the
current value of 100 is too small for any but the simplest home
networks.  Increase the size to 1000 as a compromise between the current
value and the previous default of 4096.  Simple networks that don't fill
the cache won't pay any penalty because the entries are allocated as
needed.

In the crosfleet lab, a cache size of 1000 results in CPU use of 1-2%
for avahi-daemon.  On a more typical network that doesn't have machines
constantly rebooting, it should use significantly less CPU time.  It
should be possible to reduce this further while maintaining the cache
size with some follow-up improvements to the hashmap implementation.

BUG=b:283138817
TEST=avahi-browse on several test networks

Change-Id: Id674938dfc7feb24f52306deb558b706aa4aec7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4552465
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
(cherry picked from commit e262593df6d751e7e91a9b03599af27dd380c816)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4577799
Commit-Queue: Nathan Muggli <nmuggli@google.com>
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>
Tested-by: Piotr Pawliczek <pawliczek@chromium.org>